hybrid encrypt for download --> OK

+++ app/views.py
@@ -346,15 +346,17 @@ def downloadDoc(request):
     if(request.session.get('validated', None) == None):
         return redirect('/home')
 
+    userId = request.session['userId']
+    username = request.session['username']
     errorMsg = []
 
     cursor = connection.cursor()
     if request.method == 'POST':
         # Getting the path of the file wanted
         docId = request.POST.get('docId')
-        queryString = "SELECT server_path FROM file WHERE id = '" + docId + "';"
+        queryString = "SELECT name FROM file WHERE id = '" + docId + "';"
         cursor.execute(queryString);
-        filepath = cursor.fetchone()[0];
+        filename = cursor.fetchone()[0];
 
 
         # Getting the passphrase entered y the user
@@ -364,24 +366,71 @@ def downloadDoc(request):
             errorMsg.append('You have to enter your passphrase to download your files !')
         else:
             # Checking is the passphrase is right
+            queryString = "SELECT password FROM user WHERE id = '" + str(userId) + "'"
+            cursor.execute(queryString)
+            password = cursor.fetchone()[0]
 
+            password = password.encode('utf-8')
+            IV = '-docLocker-' + username + '-' + str(len(username) * 2018)
+            IV = bytearray(hashlib.sha256(IV.encode("utf-8")).hexdigest(), 'utf-8')
+            passFilePath = 'uploaddoc/' + username + '/pass'
+            tempFileName = 'pass-temp' + str(time.time()).split('.')[0]
+            tempFilePath = 'uploaddoc/' + username + '/' + tempFileName
+            cbc('decrypt', password, IV, passFilePath, tempFilePath)
 
-            # Sending the file to the user
-            file_path = os.path.join(settings.MEDIA_ROOT, filepath)
-            if os.path.exists(file_path):
-                with open(file_path, 'rb') as f:
-                    filedata = f.read()
+            f = open(tempFilePath, 'r')
+            passUncrypted = f.read()
+            f.close()
+            os.remove(tempFilePath)
+            
+            if passUncrypted != passphrase:
+                errorMsg.append('The passphrase is incorrect')
+            else:
+                # Using the passphrase to generate RSA keys of the user
+                passphrase = hashlib.sha256(passphrase.encode("utf-8")).hexdigest()
+                salt = '-docLocker-' + username + '-' + str(userId * 2018) + '-' + str(len(username) * 2018)
+                salt = bytes(hashlib.sha256(salt.encode("utf-8")).hexdigest(), 'utf-8')   # replace with random salt if you can store one
+    
+                master_key = PBKDF2(passphrase, salt, count = 1000)  # bigger count = better
 
-                response = HttpResponse(filedata, content_type="application/vnd.ms-excel")
-                response['Content-Disposition'] = 'inline; filename=' + os.path.basename(file_path)
-                os.remove(file_path)
-                return response
-            raise Http404
+                def my_rand(n):
+                    # kluge: use PBKDF2 with count=1 and incrementing salt as deterministic PRNG
+                    my_rand.counter += 1
+                    return PBKDF2(master_key, bytes("my_rand:%d" % my_rand.counter, 'utf-8'), dkLen = n, count = 1)
+
+                my_rand.counter = 0
+                RSAkey = RSA.generate(2048, randfunc=my_rand)
+
+                keyPath = 'uploaddoc/' + username + '/key-' + str(docId) + '.bin'
+                
+                f = open(keyPath, 'rb')
+                enc_session_key, nonce, tag, ciphertext = \
+                    [ f.read(x) for x in (RSAkey.size_in_bytes(), 16, 16, -1) ]
+
+                cipher = PKCS1_OAEP.new(RSAkey)
+                session_key = cipher.decrypt(enc_session_key)
+                f.close()
+
+                IV = username + '-' + str(userId * 2018)
+                IV = bytearray(hashlib.sha256(IV.encode("utf-8")).hexdigest(), 'utf-8')
+                filepath = 'uploaddoc/' + username + '/' + filename
+                uncryptedFilePath = 'uploaddoc/' + username + '/' + filename.split('.')[0] + '-uncrypted.' + filename.split('.')[1]
+                cbc('decrypt', session_key, IV, filepath, uncryptedFilePath)
+
+                # Sending the file to the user
+                file_path = os.path.join(settings.MEDIA_ROOT, uncryptedFilePath)
+                if os.path.exists(file_path):
+                    with open(file_path, 'rb') as f:
+                        filedata = f.read()
+
+                    response = HttpResponse(filedata, content_type="application/vnd.ms-excel")
+                    response['Content-Disposition'] = 'inline; filename=' + os.path.basename(file_path)
+                    os.remove(file_path)
+                    return response
+                raise Http404
 
 
     # Querying for user's files
-    userId = request.session['userId']
-
     queryString = "SELECT name, id FROM file WHERE userId = '" + str(userId) + "';"
     cursor.execute(queryString);
     filenames = cursor.fetchmany(50);
